NewElementWindow

+++ UI/UIModels/UiMaterial.cs
@@ -6,17 +6,17 @@
         public string Naimenovanie { get; set; }
         public string Formula { get; set; }
 
-        public decimal MolekuliarniyVes { get; set; }
-        public decimal Tkipeniya { get; set; }
-		public decimal Tplavleniya { get; set; }
-		public decimal Plotnost { get; set; }
-		public decimal Viazkost { get; set; }
-		public decimal Letuchest { get; set; }
-		public decimal Cmax { get; set; }
-		public decimal RastvorimostVVode { get; set; }
-		public decimal Rastvorimost { get; set; }
-		public decimal PH { get; set; }
-		public decimal PovNatiazhenie { get; set; }
+        public float MolekuliarniyVes { get; set; }
+        public float Tkipeniya { get; set; }
+		public float Tplavleniya { get; set; }
+		public float Plotnost { get; set; }
+		public float Viazkost { get; set; }
+		public float Letuchest { get; set; }
+		public float Cmax { get; set; }
+		public float RastvorimostVVode { get; set; }
+		public float Rastvorimost { get; set; }
+		public float PH { get; set; }
+		public float PovNatiazhenie { get; set; }
 
 		public bool? Teratogennost { get; set; }
 		public bool? Kantserogennost { get; set; }
@@ -24,8 +24,8 @@
 		public bool? Imunnogennost { get; set; }
 		public bool? Gepatotoksichnost { get; set; }
 
-		public decimal LD50Zheludok { get; set; }
-		public decimal LD50kozha { get; set; }
-		public decimal LCt50 { get; set; }
+		public float LD50Zheludok { get; set; }
+		public float LD50kozha { get; set; }
+		public float LCt50 { get; set; }
     }
 }
